Send response as array for security reasons, and treat "value" as a float.

+++ backend/app/Http/Requests/Sale/StoreRequest.php
@@ -33,7 +33,7 @@ class StoreRequest extends FormRequest
         return $this->validated('seller_id');
     }
 
-    public function value(): int
+    public function value(): float
     {
         return $this->validated('value');
     }
